+++ Microsoft.PowerShell_profile.ps1
@@ -143,6 +143,7 @@ $_pf_components = @(
     "components\system\fonts.ps1"
     "components\system\login.ps1"
     "components\system\sysconfig.ps1"
+    "components\system\startup.ps1"
 
     "components\network\servers.ps1"
 )